web: address review comments
