+++ web/public/assets/js/app/__tests__/message-replies.test.js
@@ -156,6 +156,11 @@ test('buildMessageBody appends reaction counts for REACTION_APP packets without 
 // buildMessageBody — renderMentionHtml callback
 // ---------------------------------------------------------------------------
 
+// Shared mock helpers reused across the mention-callback tests below.
+const esc = v => `ESC(${v})`;
+const emoji = v => `EMOJI(${v})`;
+const badge = name => `BADGE(${name})`;
+
 test('buildMessageBody throws TypeError when renderMentionHtml is not a function', () => {
   assert.throws(
     () => buildMessageBody({
@@ -171,8 +176,8 @@ test('buildMessageBody throws TypeError when renderMentionHtml is not a function
 test('buildMessageBody without renderMentionHtml escapes @[Name] literally', () => {
   const body = buildMessageBody({
     message: { text: 'hello @[Alice]' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
   });
   assert.equal(body, 'ESC(hello @[Alice])');
 });
@@ -180,9 +185,9 @@ test('buildMessageBody without renderMentionHtml escapes @[Name] literally', () 
 test('buildMessageBody with renderMentionHtml replaces single @[Name] mention', () => {
   const body = buildMessageBody({
     message: { text: 'hi @[Alice] there' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
-    renderMentionHtml: name => `BADGE(${name})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
+    renderMentionHtml: badge,
   });
   assert.equal(body, 'ESC(hi )BADGE(Alice)ESC( there)');
 });
@@ -191,8 +196,8 @@ test('buildMessageBody with renderMentionHtml handles multiple mentions', () => 
   const calls = [];
   const body = buildMessageBody({
     message: { text: '@[A] and @[B]' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
     renderMentionHtml: (name) => { calls.push(name); return `BADGE(${name})`; },
   });
   assert.deepEqual(calls, ['A', 'B']);
@@ -203,8 +208,8 @@ test('buildMessageBody with renderMentionHtml escapes literal segments', () => {
   const body = buildMessageBody({
     message: { text: '<b> @[Alice]' },
     escapeHtml: v => v.replace(/</g, '&lt;').replace(/>/g, '&gt;'),
-    renderEmojiHtml: v => `EMOJI(${v})`,
-    renderMentionHtml: name => `BADGE(${name})`,
+    renderEmojiHtml: emoji,
+    renderMentionHtml: badge,
   });
   assert.equal(body, '&lt;b&gt; BADGE(Alice)');
 });
@@ -212,9 +217,9 @@ test('buildMessageBody with renderMentionHtml escapes literal segments', () => {
 test('buildMessageBody with renderMentionHtml at start of text', () => {
   const body = buildMessageBody({
     message: { text: '@[Alice] hello' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
-    renderMentionHtml: name => `BADGE(${name})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
+    renderMentionHtml: badge,
   });
   assert.equal(body, 'BADGE(Alice)ESC( hello)');
 });
@@ -222,9 +227,9 @@ test('buildMessageBody with renderMentionHtml at start of text', () => {
 test('buildMessageBody with renderMentionHtml at end of text', () => {
   const body = buildMessageBody({
     message: { text: 'hello @[Alice]' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
-    renderMentionHtml: name => `BADGE(${name})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
+    renderMentionHtml: badge,
   });
   assert.equal(body, 'ESC(hello )BADGE(Alice)');
 });
@@ -233,8 +238,8 @@ test('buildMessageBody with renderMentionHtml: no mentions, callback not invoked
   let called = false;
   const body = buildMessageBody({
     message: { text: 'plain text' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
     renderMentionHtml: () => { called = true; return 'BADGE'; },
   });
   assert.equal(called, false);
@@ -244,8 +249,8 @@ test('buildMessageBody with renderMentionHtml: no mentions, callback not invoked
 test('buildMessageBody with renderMentionHtml: null renderMentionHtml behaves like no callback', () => {
   const body = buildMessageBody({
     message: { text: 'hi @[Alice]' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
     renderMentionHtml: null,
   });
   assert.equal(body, 'ESC(hi @[Alice])');
@@ -256,8 +261,8 @@ test('buildMessageBody reaction path unaffected by renderMentionHtml', () => {
   let called = false;
   const body = buildMessageBody({
     message: reaction,
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
     renderMentionHtml: () => { called = true; return 'BADGE'; },
   });
   assert.equal(called, false);
@@ -267,8 +272,8 @@ test('buildMessageBody reaction path unaffected by renderMentionHtml', () => {
 test('buildMessageBody with renderMentionHtml: unclosed @[ treated as literal', () => {
   const body = buildMessageBody({
     message: { text: 'hello @[unclosed' },
-    escapeHtml: v => `ESC(${v})`,
-    renderEmojiHtml: v => `EMOJI(${v})`,
+    escapeHtml: esc,
+    renderEmojiHtml: emoji,
     renderMentionHtml: () => 'BADGE',
   });
   // @[ without closing ] does not match the pattern — treated as literal
