+++ web/public/assets/js/app/main.js
@@ -2003,7 +2003,22 @@ export function initializeApp(config) {
     if (!short) {
       return `<span class="short-name" style="background:#ccc"${titleAttr}${infoAttr}>?&nbsp;&nbsp;&nbsp;</span>`;
     }
-    const padded = escapeHtml(String(short).padStart(4, ' ')).replace(/ /g, '&nbsp;');
+    // Centre the label within a 4-column badge.  padStart alone only adds
+    // leading spaces, producing "   C" for a 1-char name with no trailing
+    // space.  Instead distribute padding evenly: 1-char → "  C ", 2-char →
+    // " AB ", 3-char → " ABC", 4-char → unchanged.  Names already at 4+
+    // chars are left as-is (meshtastic always stores exactly 4; the Ruby
+    // COMPANION override also produces exactly 4).
+    const raw = String(short);
+    let centred;
+    if (raw.length >= 4) {
+      centred = raw;
+    } else {
+      const leading = Math.ceil((4 - raw.length) / 2);
+      const trailing = 4 - raw.length - leading;
+      centred = ' '.repeat(leading) + raw + ' '.repeat(trailing);
+    }
+    const padded = escapeHtml(centred).replace(/ /g, '&nbsp;');
     const protocol = nodeData?.protocol ?? null;
     const color = getRoleColor(roleValue, protocol);
     const textColor = getRoleTextColor(roleValue, protocol);
